Shortcut dialog (#171)
Using update shortcut dialog adds another complexity to the code. I just merged the logic of add and update into Shortcut dialog instead and reduces a lot of lines of code. I also improved the snackbar tests to assert if it displays a correct message rather than asserting if snackbar is displayed or not.